advertisementDToo

+++ src/main/java/com/secondlife/demo/repository/AdvertisementRepository.java
@@ -3,5 +3,5 @@ package com.secondlife.demo.repository;
 import com.secondlife.demo.model.Advertisement;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdvertisementRepository extends JpaRepository<Advertisement, Integer> {
+public interface AdvertisementRepository extends JpaRepository<Advertisement, Long> {
 }
